# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "BAYT ALMOSWER",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
...
        "/v1/leads/{lead}/transition": {
            "patch": {
                "operationId": "lead.transition",
                "summary": "\u062a\u063a\u064a\u064a\u0631 \u0627\u0644\u0645\u0631\u062d\u0644\u0629 \u0648\u0641\u0642 \u0642\u0648\u0627\u0639\u062f Pipeline. PRD \u00a713",
                "tags": [
                    "Lead"
                ],
                "parameters": [
                    {
                        "name": "lead",
                        "in": "path",
                        "required": true,
                        "description": "The lead ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransitionLeadRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`LeadResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LeadResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
        "/v1/leads/{lead}/interactions": {
            "post": {
                "operationId": "lead.addInteraction",
                "summary": "\u062a\u0633\u062c\u064a\u0644 \u062a\u0641\u0627\u0639\u0644 \u0645\u0639 Lead. PRD \u00a713",
                "tags": [
                    "Lead"
                ],
                "parameters": [
                    {
                        "name": "lead",
                        "in": "path",
                        "required": true,
                        "description": "The lead ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "call",
                                            "whatsapp",
                                            "sms",
                                            "email",
                                            "visit",
                                            "form",
                                            "social",
                                            "note"
                                        ]
                                    },
                                    "body": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "maxLength": 5000
                                    },
                                    "occurred_at": {
                                        "type": [
                                            "string",
                                            "null"
                                        ],
                                        "format": "date-time"
                                    }
                                },
                                "required": [
                                    "type"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "const": "\u062a\u0645 \u062a\u0633\u062c\u064a\u0644 \u0627\u0644\u062a\u0641\u0627\u0639\u0644."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "LeadResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "full_name": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "phone_e164": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "email": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "stage": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "lost_reason": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "owner_user_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "person_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "lead_source_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "campaign_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "next_follow_up_at": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "full_name",
                    "phone_e164",
                    "email",
                    "stage",
                    "status",
                    "lost_reason",
                    "owner_user_id",
                    "person_id",
                    "lead_source_id",
                    "campaign_id",
                    "next_follow_up_at",
                    "created_at"
                ],
                "title": "LeadResource"
            },
...
            "TransitionLeadRequest": {
                "type": "object",
                "description": "\u062a\u062d\u0642\u0642 \u062a\u063a\u064a\u064a\u0631 \u0645\u0631\u062d\u0644\u0629 Lead. PRD \u00a713.\n\u0642\u0648\u0627\u0639\u062f: \u0627\u0644\u0625\u063a\u0644\u0627\u0642 \u0643\u0645\u0641\u0642\u0648\u062f (lost) \u064a\u062a\u0637\u0644\u0628 \u0633\u0628\u0628\u064b\u0627\u061b \u0645\u0631\u0627\u062d\u0644 \u0627\u0644\u0646\u0638\u0627\u0645 \u0644\u0627 \u062a\u064f\u0636\u0628\u0637 \u064a\u062f\u0648\u064a\u064b\u0627.",
                "properties": {
                    "stage": {
                        "type": "string",
                        "enum": [
                            "new",
                            "assigned",
                            "contacted",
                            "qualified",
                            "interested",
                            "payment_pending",
                            "nurturing",
                            "lost"
                        ]
                    },
                    "lost_reason": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 255
                    }
                },
                "required": [
                    "stage"
                ],
                "title": "TransitionLeadRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}